requirements.txt
Flask-RESTf ul==0.3.8
PyJWT==1.7.1
Flask-SQLAlchemy==2.4.1